+++ client/app/account/components/index.ts
@@ -1,6 +1,6 @@
 export { AccountInfo } from './AccountInfo'
 export { AdminWarningModal } from './AdminWarningModal'
+export { ChangePasswordForm } from './ChangePasswordForm'
 export { DeleteAccount } from './DeleteAccount'
 export { DeleteAccountModal } from './DeleteAccountModal'
-export { ChangePasswordForm } from './ChangePasswordForm'
 export { UpdateAccountForm } from './UpdateAccountForm'
