+++ server/src/user/dto/index.ts
@@ -1,0 +1,2 @@
+export * from './change-password-dto';
+export * from './update-user-dto';
